requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
requests
langchain
langchain-community
langchain-huggingface
langchain-groq
langgraph
duckduckgo-search
faiss-cpu
sentence-transformers

